fix float bug

+++ src/main/java/onefengma/demo/server/model/apibeans/product/CarProductBrief.java
@@ -1,5 +1,7 @@
 package onefengma.demo.server.model.apibeans.product;
 
+import java.math.BigDecimal;
+
 /**
  * Created by chufengma on 16/7/6.
  */
@@ -9,7 +11,7 @@ public class CarProductBrief {
     public String cover;
     public String desc;
     public int productType;
-    public float price;
+    public BigDecimal price;
     public String sourceCity;
     public String unit;
 }
