+++ src/main/java/onefengma/demo/server/model/product/HandingProductBrief.java
@@ -1,5 +1,7 @@
 package onefengma.demo.server.model.product;
 
+import java.math.BigDecimal;
+
 /**
  * Created by chufengma on 16/6/25.
  */
@@ -8,10 +10,10 @@ public class HandingProductBrief {
     public String type;
     public String souCityId;
     public String title;
-    public float price;
+    public BigDecimal price;
     public String unit;
     public String cover;
-    public float monthSellCount;
+    public BigDecimal monthSellCount;
     public float score;
 
     private String sourceCity;
